requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.2.5
packaging==25.0
pandas==2.2.3
PyJWT==2.10.1
pymongo==4.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.40
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
